Adapted internal calls to restructured negfc and metrics routines now placed in 'fm'
