+++ vip_hci/fm/negfc_fmerit.py
@@ -10,7 +10,7 @@ __all__ = []
 import numpy as np
 from hciplot import plot_frames
 from skimage.draw import disk
-from ..metrics import cube_inject_companions
+from ..fm import cube_inject_companions
 from ..var import (frame_center, get_annular_wedge, cube_filter_highpass, dist,
                    get_circle)
 from ..psfsub import pca_annulus, pca_annular, pca
