+++ vip_hci/fm/utils_negfc.py
@@ -9,7 +9,7 @@ __author__ = 'Carlos Alberto Gomez Gonzalez'
 __all__ = ['cube_planet_free']
 
 import numpy as np
-from ..metrics import cube_inject_companions
+from ..fm import cube_inject_companions
 
 
 def cube_planet_free(planet_parameter, cube, angs, psfn, plsc, imlib='vip-fft',
